+++ internal/test/immutable/oneof_test.go
@@ -1,0 +1,32 @@
+package immutable_test
+
+// import (
+// 	"testing"
+
+// 	. "github.com/dogmatiq/primo/internal/test/mutator"
+// )
+
+// func TestOneOf(t *testing.T) {
+// 	t.Parallel()
+
+// 	testInterfaces[*OneOf](
+// 		t,
+// 		MutableOneOf.SetFieldA,
+// 		ImmutableOneOf.GetFieldA,
+// 		123,
+// 	)
+
+// 	testInterfaces[*OneOf](
+// 		t,
+// 		MutableOneOf.SetFieldB,
+// 		ImmutableOneOf.GetFieldB,
+// 		456,
+// 	)
+
+// 	testInterfaces[*OneOf](
+// 		t,
+// 		MutableOneOf.SetFieldC,
+// 		ImmutableOneOf.GetFieldC,
+// 		"<value>",
+// 	)
+// }